Move reading and writing Activity to the interceptor/dispatcher implementation
