+++ src/IceRpc/Endpoint.cs
@@ -3,7 +3,6 @@
 using IceRpc.Internal;
 using Microsoft.Extensions.Logging;
 using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Globalization;
