+++ src/IceRpc/IncomingRequest.cs
@@ -5,10 +5,6 @@ using IceRpc.Interop;
 using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
-using System.Diagnostics;
-using System.Linq;
-using System.Net.Http.Headers;
-using System.Web;
 
 namespace IceRpc
 {
@@ -245,46 +241,6 @@ namespace IceRpc
 
             Payload = request.Payload.AsArraySegment();
             PayloadCompressionFormat = request.PayloadCompressionFormat;
-        }
-
-        internal void RestoreActivityContext(Activity activity)
-        {
-            Debug.Assert(Protocol == Protocol.Ice2);
-            if (BinaryContext.TryGetValue((int)BinaryContextKey.TraceContext, out ReadOnlyMemory<byte> buffer))
-            {
-                // Read W3C traceparent binary encoding (1 byte version, 16 bytes trace Id, 8 bytes span Id,
-                // 1 byte flags) https://www.w3.org/TR/trace-context/#traceparent-header-field-values
-                int i = 0;
-                byte traceIdVersion = buffer.Span[i++];
-                var traceId = ActivityTraceId.CreateFromBytes(buffer.Span.Slice(i, 16));
-                i += 16;
-                var spanId = ActivitySpanId.CreateFromBytes(buffer.Span.Slice(i, 8));
-                i += 8;
-                var traceFlags = (ActivityTraceFlags)buffer.Span[i++];
-
-                activity.SetParentId(traceId, spanId, traceFlags);
-
-                // Read tracestate encoded as a string
-                var istr = new InputStream(buffer[i..], Encoding.V20);
-                activity.TraceStateString = istr.ReadString();
-
-                // The min element size is 2 bytes for a struct with two empty strings.
-                IEnumerable<(string key, string value)> baggage = istr.ReadSequence(
-                    minElementSize: 2,
-                    istr =>
-                    {
-                        string key = istr.ReadString();
-                        string value = istr.ReadString();
-                        return (key, value);
-                    });
-
-                // Restore in reverse order to keep the order in witch the peer add baggage entries,
-                // this is important when there are duplicate keys.
-                foreach ((string key, string value) in baggage.Reverse())
-                {
-                    activity.AddBaggage(key, value);
-                }
-            }
         }
     }
 }
