+++ src/IceRpc/Interceptors-Telemetry.cs
@@ -1,7 +1,6 @@
 // Copyright (c) ZeroC, Inc. All rights reserved.
 
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Logging.Abstractions;
 using System;
 using System.Diagnostics;
 
@@ -27,51 +26,9 @@ namespace IceRpc
         /// <summary>An interceptor that start an <see cref="Activity"/> per request, following OpenTelemetry
         /// conventions. The Activity is started if the ActivitySource has any active listeners,
         /// if <see cref="Activity.Current"/> is not null or if IceRpc logger is enabled.</summary>
-        /// <param name="tracerOptions">Options to configure the tracer interceptor.</param>
+        /// <param name="options">Options to configure the tracer interceptor.</param>
         /// <returns>The CustomTracer interceptor.</returns>
-        public static Func<IInvoker, IInvoker> CustomTelemetry(TelemetryOptions tracerOptions)
-        {
-            ILogger logger = (tracerOptions.LoggerFactory ?? NullLoggerFactory.Instance).CreateLogger("IceRpc");
-
-            return next => new InlineInvoker(
-                async (request, cancel) =>
-                {
-                    if (request.Protocol == Protocol.Ice2)
-                    {
-                        Activity? activity = tracerOptions.ActivitySource?.StartActivity(
-                            $"{request.Path}/{request.Operation}",
-                            ActivityKind.Client);
-                        if (activity == null && (logger.IsEnabled(LogLevel.Critical) || Activity.Current != null))
-                        {
-                            activity = new Activity($"{request.Path}/{request.Operation}");
-                            activity.Start();
-                        }
-
-                        if (activity != null)
-                        {
-                            activity.AddTag("rpc.system", "icerpc");
-                            activity.AddTag("rpc.service", request.Path);
-                            activity.AddTag("rpc.method", request.Operation);
-                            // TODO add additional attributes
-                            // https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/trace/semantic_conventions/rpc.md#common-remote-procedure-call-conventions
-                        }
-
-                        request.WriteActivityContext();
-
-                        try
-                        {
-                            return await next.InvokeAsync(request, cancel).ConfigureAwait(false);
-                        }
-                        finally
-                        {
-                            activity?.Stop();
-                        }
-                    }
-                    else
-                    {
-                        return await next.InvokeAsync(request, cancel).ConfigureAwait(false);
-                    }
-                });
-        }
+        public static Func<IInvoker, IInvoker> CustomTelemetry(TelemetryOptions options) =>
+            next => new Internal.TelemetryInvoker(options, next);
     }
 }
