+++ src/IceRpc/OutgoingRequest.cs
@@ -6,8 +6,6 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Diagnostics;
-using System.Net.Http.Headers;
-using System.Web;
 
 namespace IceRpc
 {
@@ -261,73 +259,6 @@ namespace IceRpc
 
         /// <inheritdoc/>
         internal override IncomingFrame ToIncoming() => new IncomingRequest(this);
-
-        internal void WriteActivityContext()
-        {
-            Debug.Assert(Protocol == Protocol.Ice2);
-            if (Activity.Current is Activity activity && activity.Id != null)
-            {
-                if (activity.IdFormat != ActivityIdFormat.W3C)
-                {
-                    throw new ArgumentException("only W3C ID format is supported with IceRpc", nameof(activity));
-                }
-
-                if (activity.Id == null)
-                {
-                    throw new ArgumentException("invalid null activity ID", nameof(activity));
-                }
-
-                // For Ice2 the activity context is written to the binary context as if it has the following Slice
-                // definition
-                //
-                // struct BaggageEntry
-                // {
-                //    string key;
-                //    string value;
-                // }
-                // sequence<BaggageEntry> Baggage;
-                //
-                // struct ActivityContext
-                // {
-                //    // ActivityID version 1 byte
-                //    byte version;
-                //    // ActivityTraceId 16 bytes
-                //    ulong activityTraceId0;
-                //    ulong activityTraceId1;
-                //    // ActivitySpanId 8 bytes
-                //    ulong activitySpanId
-                //    // ActivityTraceFlags 1 byte
-                //    byte ActivityTraceFlags;
-                //    string traceStateString;
-                //    Baggage baggage;
-                // }
-
-                BinaryContextOverride.Add(
-                    (int)BinaryContextKey.TraceContext,
-                    ostr =>
-                    {
-                        // W3C traceparent binary encoding (1 byte version, 16 bytes trace Id, 8 bytes span Id,
-                        // 1 byte flags) https://www.w3.org/TR/trace-context/#traceparent-header-field-values
-                        ostr.WriteByte(0);
-                        Span<byte> buffer = stackalloc byte[16];
-                        activity.TraceId.CopyTo(buffer);
-                        ostr.WriteByteSpan(buffer);
-                        activity.SpanId.CopyTo(buffer[0..8]);
-                        ostr.WriteByteSpan(buffer[0..8]);
-                        ostr.WriteByte((byte)activity.ActivityTraceFlags);
-
-                        // Tracestate encoded as an string
-                        ostr.WriteString(activity.TraceStateString ?? "");
-
-                        // Baggage encoded as a sequence<BaggageEntry>
-                        ostr.WriteSequence(activity.Baggage, (ostr, entry) =>
-                        {
-                            ostr.WriteString(entry.Key);
-                            ostr.WriteString(entry.Value ?? "");
-                        });
-                    });
-            }
-        }
 
         /// <inheritdoc/>
         internal override void WriteHeader(OutputStream ostr)
